Developed hangman-game
The game is developed and tested. It is working fine on the website. User have to guess the word. If user guess incorrect then parts of a hangman figure appears and if the hangman figure is fully formed then user loses. User can use the keyboard appearing on the screen or he/she can his own keyboard.