wsBcBAABCAAQBQJlHzN8CRBK7hj4Ov3rIwAAnAsIAKuVupWLqqb+8bdXBqAtL1Qz  qo5u1XBwYBLcdJToxzjxgqZYHDdAITJ34p8xg00k6yCQW+OOwibmvOGwEkrKBCky  agAYS74R28v51VaOQm1S9vOw1/kTvxuiAWhMfXKCaNAOWU6KTVDgDngv3AIo49kv  9e8GG+56/s2Cv7l7NVtWSjNlxZgRh5UeNIv8Oi/cVyKxx/X260yL7kvu/yiN0ru2  jHvaeODHw1dma1Aw3La/00kUPrjo5WV8Mz3aSvyPi0M3sivKkbdXdeVUn4wV3buf  CbizRfWKmHeuUR0AeKpNamKxl0J8heA1ALRYIgIcNOVO34/TVIEhsBbRNOj88Tw=  =8/kJ  -----END PGP SIGNATURE-----
Hangman game (#683)

* Developed hangman-game

The game is developed and tested. It is working fine on the website. User have to guess the word. If user guess incorrect then parts of a hangman figure appears and if the hangman figure is fully formed then user loses. User can use the keyboard appearing on the screen or he/she can his own keyboard.

* Developed hangman-game and converted components to block components

* Developed hangman-game and converted components to block components