fix: `log` compiler option
There was a confusing behavior caused by the strange logic in the compiler
logging. It is possible to turn off the logging only if the `log` option has
been set to a truthy value, which is not equal to "info", because of the
condition `!ctx.options.log || ctx.options.log === 'info'`. However, at the same
time, the type of the `log` option is either `boolean | 'info' | undefined`.
I think that this PR will make the logic of the `log` option clearer.